* GNU Build system is newly introduced for YLISP. * Refining code to support 64bit host. * Policy of coding syntax is changed. (GCC only)     GCC specific syntax is used.

+++ ylisp/yldef.h
@@ -40,7 +40,7 @@
 #endif
 
 #ifndef offset_of
-#   define offset_of(type, member) ((unsigned int) &((type*)0)->member)
+#   define offset_of(type, member) ((unsigned long) &((type*)0)->member)
 #endif
 
 #ifndef container_of
